s: move sys to separate file
To reduce indentation and improve separation.
