+++ crates/ffi_patterns/src/lib.rs
@@ -1,2 +1,4 @@
 #[cfg(test)]
+mod sys;
+#[cfg(test)]
 mod tests;
